+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 - Validate donor and sample IDs.
 - Check that CLI docs are up to date.
 - Validate the data_path.
+- Allow multiple metadata.tsvs.
+- Validate against globus.
 ### Changed
 - CSV -> TSV
 - Make the schema validation errors more readable